Move sitemap prop fetching logic into `page-props`
