+++ src/adapter-nextjs/page-props/index.ts
@@ -3,5 +3,6 @@ export { getGalleryPageServerSideProps, type GalleryPageProps } from './gallery'
 export { getGalleryAlbumPageServerSideProps, type GalleryAlbumPageProps } from './galleryAlbum';
 export { getHomepageServerSideProps, type HomePageProps } from './home';
 export { getSearchPageServerSideProps } from './search';
+export * from './sitemap';
 export { getStoryPageServerSideProps } from './story';
 export { getStoryPreviewPageServerSideProps } from './storyPreview';
