+++ src/adapter-nextjs/page-props/sitemap/index.ts
@@ -1,4 +1,4 @@
-export { getInitialProps as getSitemapInitialProps } from './getInitialProps';
+export { getSitemapServerSideProps } from './getServerSideProps';
 
 export { SitemapBuilder } from './SitemapBuilder';
 export { createPaths } from './createPaths';
