+++ src/components-nextjs/index.ts
@@ -1,3 +1,2 @@
 export { PageSeo } from './PageSeo';
-export * from './Sitemap';
 export { StorySeo } from './StorySeo';
